# Improve BatchStreamReader - Part 1