--- a/presto-orc/src/main/java/com/facebook/presto/orc/AbstractOrcRecordReader.java
+++ b/presto-orc/src/main/java/com/facebook/presto/orc/AbstractOrcRecordReader.java
@@ -348,6 +348,11 @@ abstract class AbstractOrcRecordReader<T extends StreamReader>
     {
         try (Closer closer = Closer.create()) {
             closer.register(orcDataSource);
+            for (StreamReader column : streamReaders) {
+                if (column != null) {
+                    closer.register(column::close);
+                }
+            }
         }
 
         if (writeChecksumBuilder.isPresent()) {