--- a/presto-orc/src/main/java/com/facebook/presto/orc/OrcBatchRecordReader.java
+++ b/presto-orc/src/main/java/com/facebook/presto/orc/OrcBatchRecordReader.java
@@ -67,6 +67,8 @@ public class OrcBatchRecordReader
             Optional<OrcWriteValidation> writeValidation,
             int initialBatchSize,
             StripeMetadataSource stripeMetadataSource)
+            throws OrcCorruptionException
+
     {
         super(includedColumns,
                 // The streamReadersSystemMemoryContext covers the StreamReader local buffer sizes, plus leaf node StreamReaders'
@@ -120,10 +122,10 @@ public class OrcBatchRecordReader
         return batchSize;
     }
 
-    public Block readBlock(Type type, int columnIndex)
+    public Block readBlock(int columnIndex)
             throws IOException
     {
-        Block block = getStreamReaders()[columnIndex].readBlock(type);
+        Block block = getStreamReaders()[columnIndex].readBlock();
         updateMaxCombinedBytesPerRow(columnIndex, block);
         return block;
     }
@@ -144,7 +146,7 @@ public class OrcBatchRecordReader
         if (shouldValidateWritePageChecksum()) {
             Block[] blocks = new Block[getStreamReaders().length];
             for (int columnIndex = 0; columnIndex < getStreamReaders().length; columnIndex++) {
-                blocks[columnIndex] = readBlock(includedColumns.get(columnIndex), columnIndex);
+                blocks[columnIndex] = readBlock(columnIndex);
             }
             Page page = new Page(batchSize, blocks);
             validateWritePageChecksum(page);
@@ -156,6 +158,7 @@ public class OrcBatchRecordReader
             List<OrcType> types,
             DateTimeZone hiveStorageTimeZone,
             Map<Integer, Type> includedColumns)
+            throws OrcCorruptionException
     {
         List<StreamDescriptor> streamDescriptors = createStreamDescriptor("", "", 0, types, orcDataSource).getNestedStreams();
 
@@ -163,8 +166,11 @@ public class OrcBatchRecordReader
         BatchStreamReader[] streamReaders = new BatchStreamReader[rowType.getFieldCount()];
         for (int columnId = 0; columnId < rowType.getFieldCount(); columnId++) {
             if (includedColumns.containsKey(columnId)) {
-                StreamDescriptor streamDescriptor = streamDescriptors.get(columnId);
-                streamReaders[columnId] = BatchStreamReaders.createStreamReader(streamDescriptor, hiveStorageTimeZone);
+                Type type = includedColumns.get(columnId);
+                if (type != null) {
+                    StreamDescriptor streamDescriptor = streamDescriptors.get(columnId);
+                    streamReaders[columnId] = BatchStreamReaders.createStreamReader(type, streamDescriptor, hiveStorageTimeZone);
+                }
             }
         }
         return streamReaders;