Cherry-pick of https://github.com/prestosql/presto/pull/555

The memory tracking for StreamReader was firstly introduced in 97552478c9 Add memory tracking for StreamReader local buffers and then removed in
303cecacf7 Remove unnecessary StreamReader.close method
47423bbb2c Remove unused systemMemoryContext from OrcBatchRecordReader
be749f6197 Remove unused systemMemoryContext variable from XxxBatchStreamReader

BenchmarkBatchStreamReaders results:

Before
Benchmark                                  (typeSignature)  (withNulls)  Mode  Cnt    Score   Error  Units
BenchmarkBatchStreamReaders.readBlocks             boolean      PARTIAL  avgt   60   45.432 ± 0.720  ms/op
BenchmarkBatchStreamReaders.readBlocks             boolean         NONE  avgt   60   34.396 ± 0.629  ms/op
BenchmarkBatchStreamReaders.readBlocks             boolean          ALL  avgt   60    7.894 ± 0.166  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint      PARTIAL  avgt   60   47.784 ± 1.106  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint         NONE  avgt   60   39.331 ± 0.747  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint          ALL  avgt   60    8.269 ± 0.217  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint      PARTIAL  avgt   60   87.624 ± 3.163  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint         NONE  avgt   60   93.390 ± 2.122  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint          ALL  avgt   60    8.886 ± 0.289  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer      PARTIAL  avgt   60   97.352 ± 2.974  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer         NONE  avgt   60  106.238 ± 2.092  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer          ALL  avgt   60    9.017 ± 0.301  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint      PARTIAL  avgt   60  111.976 ± 3.586  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint         NONE  avgt   60  133.045 ± 4.189  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint          ALL  avgt   60    8.871 ± 0.322  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp      PARTIAL  avgt   60  165.562 ± 5.700  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp         NONE  avgt   60  238.276 ± 7.139  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp          ALL  avgt   60    9.239 ± 0.506  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct      PARTIAL  avgt   60  208.502 ± 3.426  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct         NONE  avgt   60  279.045 ± 6.075  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct          ALL  avgt   60   32.427 ± 0.826  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary      PARTIAL  avgt   60   52.460 ± 1.217  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary         NONE  avgt   60   38.757 ± 0.596  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary          ALL  avgt   60   32.064 ± 0.518  ms/op

After
Benchmark                                  (typeSignature)  (withNulls)  Mode  Cnt    Score   Error  Units
BenchmarkBatchStreamReaders.readBlocks             boolean      PARTIAL  avgt   60   26.017 ± 0.436  ms/op
BenchmarkBatchStreamReaders.readBlocks             boolean         NONE  avgt   60   14.980 ± 0.255  ms/op
BenchmarkBatchStreamReaders.readBlocks             boolean          ALL  avgt   60    8.180 ± 0.195  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint      PARTIAL  avgt   60   27.405 ± 0.604  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint         NONE  avgt   60   16.941 ± 2.228  ms/op
BenchmarkBatchStreamReaders.readBlocks             tinyint          ALL  avgt   60    7.880 ± 0.181  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint      PARTIAL  avgt   60   43.038 ± 1.083  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint         NONE  avgt   60   49.573 ± 2.417  ms/op
BenchmarkBatchStreamReaders.readBlocks            smallint          ALL  avgt   60    8.062 ± 0.105  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer      PARTIAL  avgt   60   49.302 ± 0.955  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer         NONE  avgt   60   56.010 ± 1.357  ms/op
BenchmarkBatchStreamReaders.readBlocks             integer          ALL  avgt   60    8.114 ± 0.265  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint      PARTIAL  avgt   60   60.625 ± 1.964  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint         NONE  avgt   60   82.082 ± 4.071  ms/op
BenchmarkBatchStreamReaders.readBlocks              bigint          ALL  avgt   60    8.117 ± 0.136  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp      PARTIAL  avgt   60  131.245 ± 3.605  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp         NONE  avgt   60  199.041 ± 4.008  ms/op
BenchmarkBatchStreamReaders.readBlocks           timestamp          ALL  avgt   60    8.134 ± 0.155  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct      PARTIAL  avgt   60   61.129 ± 1.331  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct         NONE  avgt   60   71.723 ± 2.255  ms/op
BenchmarkBatchStreamReaders.readBlocks      varchar_direct          ALL  avgt   60   21.339 ± 0.940  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary      PARTIAL  avgt   60   55.705 ± 1.959  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary         NONE  avgt   60   29.450 ± 0.479  ms/op
BenchmarkBatchStreamReaders.readBlocks  varchar_dictionary          ALL  avgt   60   20.122 ± 0.396  ms/op


```
== RELEASE NOTES ==

Hive Changes
* Improve BatchStreamReader performance
```