@wenleix 2019-11-01
@yingsu00 : Would you mind point to the commit/PR that removes buffers (i.e. the reverse to "Add memory tracking from StreamReader local buffers"?). Thanks!