@yingsu00 2019-11-05
> @yingsu00 : Would you mind point to the commit/PR that removes buffers (i.e. the reverse to "Add memory tracking from StreamReader local buffers"?). Thanks!

I added them to the description of the PR:
The memory tracking for StreamReader was firstly introduced in 
`97552478c9 Add memory tracking for StreamReader local buffers `
and then removed in
`303cecacf7 Remove unnecessary StreamReader.close method`
`47423bbb2c Remove unused systemMemoryContext from OrcBatchRecordReader`